EslNews.Handler behaviour (ESL News v0.1.0)

EslNews.Handler provides a standard base implementing the :cowboy_rest behaviour and URI param extraction for EslNews.Handlers

Types

switch_handler()

Specs

switch_handler() ::
  {:switch_handler, module()} | {:switch_handler, module(), any()}

:cowboy_rest.switch_handler() type isn't exported by :cowboy_rest

Callbacks

allowed_methods(req, any)

Specs

allowed_methods(:cowboy_req.req(), any()) ::
  {[binary(), ...], :cowboy_req.req(), any()}
  | {:stop, :cowboy_req.req(), any()}
  | {switch_handler(), :cowboy_req.req(), any()}

content_types_provided(req, any)

Specs

content_types_provided(:cowboy_req.req(), any()) ::
  {[{binary() | {binary(), binary(), :* | [{binary(), binary()}]}, atom()}],
   :cowboy_req.req(), any()}
  | {:stop, :cowboy_req.req(), any()}
  | {switch_handler(), :cowboy_req.req(), any()}

init(req, any)

Specs

init(:cowboy_req.req(), any()) :: {:cowboy_rest, :cowboy_req.req(), any()}

resource_exists(req, any)
(optional)

Specs

resource_exists(:cowboy_req.req(), any()) ::
  {boolean(), :cowboy_req.req(), any()}
  | {:stop, :cowboy_req.req(), any()}
  | {switch_handler(), :cowboy_req.req(), any()}

response(req, any)

Specs

response(:cowboy_req.req(), any()) :: {binary(), :cowboy_req.req(), any()}

Functions

pagination_params(request)

Specs

pagination_params(:cowboy_req.req()) :: {non_neg_integer(), non_neg_integer()}

Extract pagination params from URI query string.

Examples

iex> EslNews.Handler.pagination_params(%{qs: "page=4&per=20"}) {4, 20}

iex> EslNews.Handler.pagination_params(%{qs: "page=-3"}) {1, 10}

request_params(request, opts)

Specs

request_params(:cowboy_req.req(), [{:permit, list()}]) :: map()

Extract request params from path bindings and URI query string.

For query string params, only the permitted values will be returned, and they will not override path binding values.

Examples

iex> EslNews.Handler.request_params(%{qs: "page=1&per=10"}, permit: [:page, :per]) %{page: "1", per: "10"}

iex> EslNews.Handler.request_params(%{qs: "page=1&per=10&id=4", bindings: %{id: "1234"}}, permit: [:page]) %{id: "1234", page: "1"}

iex> EslNews.Handler.request_params(%{qs: "page=1&per=10"}, permit: []) %{}

to_integer(param, default \\ 0)

Specs

to_integer(binary(), integer()) :: integer()

Helper to convert a string param to integer