EslNews.Handlers.Ws (ESL News v0.1.0)

EslNews.Handlers.Ws provides a JSON response for the latest available Stories through websockets. Implements @behaviour for EslNews.Websocket

Types

call_result()

Specs

call_result() ::
  {:cowboy_websocket.commands(), any()}
  | {:cowboy_websocket.commands(), any(), :hibernate}

:cowboy_websocket.call_result() type isn't exported by :cowboy_websocket

Functions

init(request, state)

Specs

init(:cowboy_req.req(), any()) :: {:cowboy_websocket, :cowboy_req.req(), any()}

Upgrade a basic :cowboy_handler middleware to the :cowboy_websocket middleware

terminate(call, request, state)

Specs

terminate(any(), :cowboy_req.req(), any()) :: :ok

Logs when a connection is terminated.

websocket_handle(call, state)

Specs

websocket_handle(
  :ping | :pong | {:text | :binary | :ping | :pong, binary()},
  any()
) :: call_result()

Receives pings/data from the connection. Ignores and continues.

websocket_info(call, state)

Specs

websocket_info({:timeout, any(), :connected | :tick}, any()) :: call_result()

Deliver JSON encoded list of stories to the websocket client every time the timer expires

websocket_init(state)

Specs

websocket_init(any()) :: call_result()

Log a successful connection and dispatch the initial call to websocket_info/2 after 10ms