EslNews.Http.Client (ESL News v0.1.0)

EslNews.Http.Client fetches JSON objects from an HTTP API.

Types

option()

Specs

option() ::
  {:method, Tesla.Env.method()}
  | {:url, Tesla.Env.url()}
  | {:query, Tesla.Env.query()}
  | {:headers, Tesla.Env.headers()}
  | {:body, Tesla.Env.body()}
  | {:opts, Tesla.Env.opts()}

Options that may be passed to a request function. See request/2 for detailed descriptions.

Functions

get(client, url, opts)

Specs

get(Tesla.Env.client(), Tesla.Env.url(), [option()]) :: Tesla.Env.result()

Perform a GET request.

See request/1 or request/2 for options definition.

get("/users")
get("/users", query: [scope: "admin"])
get(client, "/users")
get(client, "/users", query: [scope: "admin"])
get(client, "/users", body: %{name: "Jon"})

get!(client, url, opts)

Specs

get!(Tesla.Env.client(), Tesla.Env.url(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a GET request.

See request!/1 or request!/2 for options definition.

get!("/users")
get!("/users", query: [scope: "admin"])
get!(client, "/users")
get!(client, "/users", query: [scope: "admin"])
get!(client, "/users", body: %{name: "Jon"})

list(list)

Specs

list(atom()) :: [non_neg_integer(), ...]

List of story IDs in order of popularity

lists()

Specs

lists() :: [atom(), ...]

Available story list names

request(client \\ %Tesla.Client{}, options)

Specs

request(Tesla.Env.client(), [option()]) :: Tesla.Env.result()

Perform a request.

Options

:method - the request method, one of [:head, :get, :delete, :trace, :options, :post, :put, :patch]

:url - either full url e.g. "http://example.com/some/path" or just "/some/path" if using Tesla.Middleware.BaseUrl

:query - a keyword list of query params, e.g. [page: 1, per_page: 100]

:headers - a keyworld list of headers, e.g. [{"content-type", "text/plain"}]

:body - depends on used middleware:

by default it can be a binary

if using e.g. JSON encoding middleware it can be a nested map

if adapter supports it it can be a Stream with any of the above

:opts - custom, per-request middleware or adapter options

Examples

ExampleApi.request(method: :get, url: "/users/path")

# use shortcut methods
ExampleApi.get("/users/1")
ExampleApi.post(client, "/users", %{name: "Jon"})

request!(client \\ %Tesla.Client{}, options)

Specs

request!(Tesla.Env.client(), [option()]) :: Tesla.Env.t() | no_return()

Perform request and raise in case of error.

This is similar to request/2 behaviour from Tesla 0.x

See request/2 for list of available options.

story(id)

Specs

story(non_neg_integer() | binary()) :: map()

Raw JSON story struct

user(username)

Specs

user(binary()) :: map()

Raw JSON user struct