EslNews.Http.SyncLists (ESL News v0.1.0)

EslNews.Http.SyncLists

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_call(msg, from, state)

GenServer.call/2 callback.

:pull_item Obtain next {list, item_id} tuple and remove it from state

handle_cast(msg, state)

GenServer.cast/2 callback.

:push_item Insert a {list, item_id} tuple into state.

:sync_list Fetch list item IDs from HTTP endpoint and keep only the first @sync_items_count entries. Save the entries to a temporary upcoming EslNews.Store.List, and then create empty stub EslNews.Store.Story for each ID before pushing the ID into the processing queue.

:sweep_list Fetch list item IDs from an upcoming list and hydrate them into EslNews.Store.Story structs. Ensure no stub structs remain by checking their type attribute and once all are set, save the upcoming list as the current list. Finally, delete the upcoming list to prepare for the next sync cycle.

handle_info(msg, state)

On every :sync_tick timeout, enqueue all lists to be updated and set a new timer.

init(state)

Specs

init(any()) :: {:ok, any()}

Initialize by dispatching the initial call to :sync_tick

pull_item()

Specs

pull_item() :: {atom(), non_neg_integer()} | nil

Dequeue an {list, item_id} tuple for processing. nil if no tuples remain.

push_item(list, item)

Specs

push_item(atom(), non_neg_integer() | String.t()) :: :ok

Enqueue a {list, item_id} tuple for processing.

start_link(opts \\ [])

Specs

start_link(GenServer.options()) :: GenServer.on_start()

sweep_list(list)

Specs

sweep_list(atom()) :: :ok

Verify that all items in an upcoming list have been fetched and convert it to the current items list.

sync_list(list)

Specs

sync_list(atom()) :: :ok

Enqueue a list name to fetch from the API as an upcoming items list.