EslNews.Logger (ESL News v0.1.0)

EslNews.Logger provides logging helpers for EslNews.Handlers

Functions

info(info)

Specs

info(any()) :: :ok | nil

Log a simple value with timestamp

request(request, state)

Specs

request(:cowboy_req.req(), any()) :: :ok | nil

Log a :cowboy_req with a timestamp, method, path and optionally the query params

resource(state)

Specs

resource([Keyword.t()]) :: :ok | nil

Log the Story in a request state with a timestamp, the Story ID and title

ws(call, state)

Specs

ws(tuple(), list()) :: :ok | nil

Log a :cowboy_websocket callback with a timestamp, frame data, and state

ws_frame(arg)

Specs

ws_frame(tuple()) :: :ok | nil

Log a :cowboy_websocket frame

ws_state(state)

Specs

ws_state(list()) :: :ok | nil

Log a :cowboy_websocket state