EslNews.Store.List (ESL News v0.1.0)

EslNews.Store.List provides a :mnesia schema and struct to store EslNews.Store.List objects, as well as helper methods to interact with the storage table.

Types

r()

Specs

r() ::
  {EslNews.Store.List, atom(), [non_neg_integer(), ...],
   non_neg_integer() | nil}

EslNews.Store.List record tuple type

t()

Specs

t() :: %EslNews.Store.List{
  id: atom(),
  items: [non_neg_integer(), ...],
  time: non_neg_integer() | nil
}

EslNews.Store.List struct type

Functions

all()

Specs

all() :: list()

List all existing EslNews.Store.List records

create(state)

Specs

create(t() | {atom(), [non_neg_integer(), ...]}) ::
  :ok | :record_exists | atom()

Persist an EslNews.Store.List in :mnesia

decode(record)

Specs

decode(r()) :: t()

Decode a :mnesia record tuple into an EslNews.Store.List struct

Examples

iex> EslNews.Store.List.decode({EslNews.Store.List, :topstories, [4,2,3,1], 54321})
%EslNews.Store.List{id: :topstories, items: [4,2,3,1], time: 54321}

delete(list)

Specs

delete(t()) :: :ok

Delete an EslNews.Store.List stored in :mnesia

encode(list)

Specs

encode(t()) :: r()

Encode an EslNews.Store.List struct into a :mnesia record tuple

Examples

iex> EslNews.Store.List.encode(%EslNews.Store.List{id: :topstories, items: [4,2,3,1], time: 54321})
{EslNews.Store.List, :topstories, [4,2,3,1], 54321}

find(id)

Specs

find(atom()) :: {:ok, t()} | {:not_found, nil}

Load an EslNews.Store.List stored in :mnesia or return :not_found

lists()

Specs

lists() :: [atom(), ...]

Available story list names

schema_attrs()

Specs

schema_attrs() :: [atom(), ...]

Attributes list for :mnesia table schema definition. :id must always be the first attribute

schema_indices(keys)

Specs

schema_indices([atom(), ...]) :: [non_neg_integer()]

Zero-based list of attribute positions to be indexed for :mnesia table schema definition. :id is indexed by default, will not be included