EslNews.Store.Story (ESL News v0.1.0)

EslNews.Store.Story provides a mnesia schema and struct to store EslNews.Store.Story objects, as well as helper methods to interact with the storage table.

Types

r()

Specs

r() ::
  {EslNews.Store.Story, non_neg_integer(), String.t() | nil, non_neg_integer(),
   integer(), non_neg_integer() | nil, String.t() | nil, String.t() | nil,
   String.t() | nil}

EslNews.Store.Story record tuple type

t()

Specs

t() :: %EslNews.Store.Story{
  by: String.t(),
  descendants: non_neg_integer(),
  id: non_neg_integer(),
  score: integer(),
  time: non_neg_integer(),
  title: String.t(),
  type: String.t(),
  url: String.t() | nil
}

EslNews.Store.Story struct type

Functions

all()

Specs

all() :: list()

All existing EslNews.Store.Story records

all(list_id)

Specs

all(atom() | [non_neg_integer(), ...]) :: list()

All existing EslNews.Store.Story records for an EslNews.Store.List or list of IDs

create(id)

Specs

create(t() | non_neg_integer()) :: :ok | :record_exists | atom()

Persist an EslNews.Store.Story in :mnesia

decode(record)

Specs

decode(r()) :: t()

Decode a :mnesia record tuple into an EslNews.Store.Story struct

Examples

iex> EslNews.Store.Story.decode({EslNews.Store.Story, 1, "Author", 0, 0, 0, "Title", "story", nil})
%EslNews.Store.Story{id: 1, by: "Author", descendants: 0, score: 0, time: 0, type: "story", title: "Title", url: nil}

delete(story)

Specs

delete(t()) :: :ok

Delete an EslNews.Store.Story stored in :mnesia

encode(story)

Specs

encode(t()) :: r()

Encode an EslNews.Store.Story struct into a :mnesia record tuple

Examples

iex> EslNews.Store.Story.encode(%EslNews.Store.Story{id: 1, type: "story", title: "Title", by: "Author"})
{EslNews.Store.Story, 1, "Author", nil, nil, nil, "Title", "story", nil}

find(id)

Specs

find(non_neg_integer()) :: {:ok, t()} | {:not_found, nil}

Load an EslNews.Store.Story stored in :mnesia or return :not_found

save(state)

Specs

save(t()) :: :ok | :not_found | atom()

Update an existing EslNews.Store.Story in :mnesia

schema_attrs()

Specs

schema_attrs() :: [atom(), ...]

Attributes list for :mnesia table schema definition. :id must always be the first attribute

schema_indices(keys)

Specs

schema_indices([atom(), ...]) :: [non_neg_integer()]

Zero-based list of attribute positions to be indexed for :mnesia table schema definition. :id is indexed by default, will not be included